refactor(rust): let armonik load the options, and the transport receive them
`armonik-transport` read `GrpcClient__*` out of the environment, which is integration with a
deployment rather than transport. It now takes the configuration it is handed and nothing
else, so a caller that keeps its settings anywhere at all can use it, and the crate no longer
knows it lives inside ArmoniK.

`ClientConfigArgs::from_env` and `ClientConfig::from_env` become a `FromEnv` extension trait
in `armonik`, together with `read_env*` and `ReadEnvError`. An extension trait because the
types stay where they are and an inherent method would have to live with them. What is left
of the transport's `utils` is one certificate verifier, which moves next to its only user.

`ConfigError` loses its `Env` variant, and `armonik` gains `EnvConfigError` and
`NewClientError`. `Client::new` returns the latter: `ConnectionError` belongs to the transport
and is `#[non_exhaustive]`, so `armonik` cannot express a reading failure through it.

`#[non_exhaustive]` comes off `ClientConfigArgs`. A foreign crate cannot write a literal for
such a struct, only `Default` and field assignment, which would let an option added here go
unread there without a word. Exhaustive, the compiler says so.

Two things the move turned up. `InvalidDuration` served seven duration fields while its
message named `GrpcClient__ConnectTimeout`, so a bad `Timeout` was reported against
`ConnectTimeout`; the variant now carries the option. And the hidden context selectors
`ConfigSnafu`, `IoSnafu`, `TlsSnafu`, `TransportSnafu` were exported for `Client::new` alone,
which no longer needs them.

Error messages name the field, `cert_pem` where they said `GrpcClient__CertPem`. The one
thing this crate still reads from the environment is `ProxySource::System`, the `*_PROXY`
convention every HTTP client obeys, which is not ArmoniK's vocabulary.

cargo test --workspace --all-features --no-fail-fast -- --skip client:: : 153 passed, 0
failed. The `client::` tests want the ArmoniK mock, which CI provides and this machine does
not. clippy --workspace --all-features --all-targets, and with -Dunused-crate-dependencies:
clean. cargo fmt --all --check and RUSTDOCFLAGS=-Dwarnings cargo doc --workspace
--all-features: clean.
